+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/cart/DirectDiscountReferenceImpl.java
@@ -1,0 +1,82 @@
+
+package com.commercetools.api.models.cart;
+
+import java.time.*;
+import java.util.*;
+
+import com.commercetools.api.models.common.ReferenceTypeId;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.*;
+
+import io.vrap.rmf.base.client.ModelBase;
+import io.vrap.rmf.base.client.utils.Generated;
+
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+
+/**
+ *  <p>Reference to a DirectDiscount.</p>
+ */
+@Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
+public class DirectDiscountReferenceImpl implements DirectDiscountReference, ModelBase {
+
+    private com.commercetools.api.models.common.ReferenceTypeId typeId;
+
+    private String id;
+
+    /**
+     * create instance with all properties
+     */
+    @JsonCreator
+    DirectDiscountReferenceImpl(@JsonProperty("id") final String id) {
+        this.id = id;
+        this.typeId = ReferenceTypeId.findEnum("direct-discount");
+    }
+
+    /**
+     * create empty instance
+     */
+    public DirectDiscountReferenceImpl() {
+        this.typeId = ReferenceTypeId.findEnum("direct-discount");
+    }
+
+    /**
+     *  <p>Type of referenced resource.</p>
+     */
+
+    public com.commercetools.api.models.common.ReferenceTypeId getTypeId() {
+        return this.typeId;
+    }
+
+    /**
+     *  <p>Unique identifier of the referenced DirectDiscount.</p>
+     */
+
+    public String getId() {
+        return this.id;
+    }
+
+    public void setId(final String id) {
+        this.id = id;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+
+        if (o == null || getClass() != o.getClass())
+            return false;
+
+        DirectDiscountReferenceImpl that = (DirectDiscountReferenceImpl) o;
+
+        return new EqualsBuilder().append(typeId, that.typeId).append(id, that.id).isEquals();
+    }
+
+    @Override
+    public int hashCode() {
+        return new HashCodeBuilder(17, 37).append(typeId).append(id).toHashCode();
+    }
+
+}