+++ changes.md
@@ -1,6 +1,13 @@
 **Api changes**
 
 <details>
+<summary>Added Enum(s)</summary>
+
+- added enum `direct-discount` to type `ReferenceTypeId`
+</details>
+
+
+<details>
 <summary>Added Property(s)</summary>
 
 - added property `shippingKey` to type `Cart`
@@ -9,5 +16,19 @@
 - added property `shippingCustomFields` to type `StagedOrder`
 - added property `shippingKey` to type `Order`
 - added property `shippingCustomFields` to type `Order`
+</details>
+
+
+<details>
+<summary>Changed Property(s)</summary>
+
+- :warning: changed property `discount` of type `DiscountedLineItemPortion` from type `CartDiscountReference` to `Reference`
+</details>
+
+
+<details>
+<summary>Added Type(s)</summary>
+
+- added type `DirectDiscountReference`
 </details>